update dmu
changelog:
- re-format command:
$RTEXT,	TYPE,	param1,	param2,	"text1\ntext2"	*
- add animasi

+++ firmware/DMU/matrix/matrix.h
@@ -21,12 +21,12 @@ void rgb_swap_buffer();
 uint32_t rgb_get_buffer(uint8_t row);
 
 void rgb_frame_clear(uint8_t color);
-void rgb_draw_pixel(int16_t x, int16_t y, uint8_t color);
-void rgb_write_constrain(int16_t x, int16_t y, char c, uint8_t color, uint8_t fonstSize,
-		int16_t xMin, int16_t xMax, int16_t yMin, int16_t yMax);
-void rgb_write(int16_t x, int16_t y, char c, uint8_t color, uint8_t fontSize);
-int16_t rgb_print_constrain(int16_t x, int16_t y, char* s, uint16_t size, uint8_t color,
-		uint8_t fontSize, int16_t xMin, int16_t xMax, int16_t yMin, int16_t yMax);
-void rgb_print(int16_t x, int16_t y, char* s, uint16_t size, uint8_t color, uint8_t fontSize);
+void rgb_draw_pixel(int32_t x, int32_t y, uint8_t color);
+void rgb_write_constrain(int32_t x, int32_t y, char c, uint8_t color, uint8_t fonstSize,
+		int32_t xMin, int32_t xMax, int32_t yMin, int32_t yMax);
+void rgb_write(int32_t x, int32_t y, char c, uint8_t color, uint8_t fontSize);
+int32_t rgb_print_constrain(int32_t x, int32_t y, char* s, uint32_t size, uint8_t color,
+		uint8_t fontSize, int32_t xMin, int32_t xMax, int32_t yMin, int32_t yMax);
+void rgb_print(int32_t x, int32_t y, char* s, uint32_t size, uint8_t color, uint8_t fontSize);
 
 #endif /* MATRIX_H_ */
